+++ crates/macros/src/lib.rs
@@ -107,7 +107,9 @@ pub fn derive_create_route(input: TokenStream) -> TokenStream {
     // Note: we don't need to mention T: Deserialize<'de> here; your trait bound
     // enforces it at use sites. If you want a nicer error, you can add a where clause.
     let expanded = quote! {
-        impl<'de> axum_helpers::CreateRoute<'de> for #name {}
+        impl<'de> axum_helpers::CreateRoute<'de> for #name
+        where
+            <#name as sql_traits::InsertSQL>::ReturnType: serde::Serialize {}
     };
 
     TokenStream::from(expanded)
@@ -123,7 +125,9 @@ pub fn derive_bulk_create_route(input: TokenStream) -> TokenStream {
     // Note: we don't need to mention T: Deserialize<'de> here; your trait bound
     // enforces it at use sites. If you want a nicer error, you can add a where clause.
     let expanded = quote! {
-        impl<'de> axum_helpers::BulkCreateRoute<'de> for #name {}
+        impl<'de> axum_helpers::BulkCreateRoute<'de> for #name
+        where
+            <#name as sql_traits::BulkInsertSQL>::ReturnType: serde::Serialize {}
     };
 
     TokenStream::from(expanded)
@@ -165,8 +169,12 @@ pub fn derive_basic_crud_routes(input: TokenStream) -> TokenStream {
     let expanded = quote! {
         impl axum_helpers::GetLatestRoute for #name {}
         impl axum_helpers::ListRecordsRoute for #name {}
-        impl<'de> axum_helpers::CreateRoute<'de> for #name {}
-        impl<'de> axum_helpers::BulkCreateRoute<'de> for #name {}
+        impl<'de> axum_helpers::CreateRoute<'de> for #name
+        where
+            <#name as sql_traits::InsertSQL>::ReturnType: serde::Serialize {}
+        impl<'de> axum_helpers::BulkCreateRoute<'de> for #name
+        where
+            <#name as sql_traits::BulkInsertSQL>::ReturnType: serde::Serialize {}
         impl axum_helpers::DeleteRoute for #name {}
     };